api/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.1
Deprecated==1.2.14
deprecation==2.1.0
exceptiongroup==1.2.0
flask==3.0.1
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==23.2
postgrest==0.15.0
pycparser==2.21
pydantic==2.5.3
pydantic-core==2.14.6
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
realtime==1.0.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.3
typing-extensions==4.9.0
urllib3==2.1.0
websockets==11.0.3
werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
